Completed splitting the connection tree from the connection tree window
