+++ mRemoteNGTests/mRemoteNGTests.csproj
@@ -101,7 +101,9 @@
     </When>
     <Otherwise>
       <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework" />
+        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework">
+          <Private>False</Private>
+        </Reference>
       </ItemGroup>
     </Otherwise>
   </Choose>
@@ -146,6 +148,7 @@
     <Compile Include="Security\CryptographyProviderFactoryTests.cs" />
     <Compile Include="Security\EncryptedSecureStringTests.cs" />
     <Compile Include="Security\LegacyRijndaelCryptographyProviderTests.cs" />
+    <Compile Include="Tree\ConnectionTreeTests.cs" />
     <Compile Include="Tree\NodeSearcherTests.cs" />
     <Compile Include="Tree\RootNodeInfoTests.cs" />
     <Compile Include="UI\Controls\TestForm.cs">
