+++ mRemoteV1/App/Runtime.cs
@@ -42,8 +42,8 @@ namespace mRemoteNG.App
         public static SecureString EncryptionKey { get; set; } = new RootNodeInfo(RootNodeType.Connection).PasswordString.ConvertToSecureString();
         public static ConnectionTreeModel ConnectionTreeModel
         {
-            get { return Windows.TreeForm.ConnectionTreeModel; }
-            set { Windows.TreeForm.ConnectionTreeModel = value; }
+            get { return Windows.TreeForm.ConnectionTree.ConnectionTreeModel; }
+            set { Windows.TreeForm.ConnectionTree.ConnectionTreeModel = value; }
         }
         #endregion
 
@@ -237,7 +237,7 @@ namespace mRemoteNG.App
                 // Load config
                 connectionsLoader.ConnectionFileName = filename;
                 ConnectionTreeModel = connectionsLoader.LoadConnections(false);
-                Windows.TreeForm.ConnectionTreeModel = ConnectionTreeModel;
+                Windows.TreeForm.ConnectionTree.ConnectionTreeModel = ConnectionTreeModel;
             }
             catch (Exception ex)
             {
@@ -288,7 +288,7 @@ namespace mRemoteNG.App
 
                 connectionsLoader.UseDatabase = Settings.Default.UseSQLServer;
                 ConnectionTreeModel = connectionsLoader.LoadConnections(false);
-                Windows.TreeForm.ConnectionTreeModel = ConnectionTreeModel;
+                Windows.TreeForm.ConnectionTree.ConnectionTreeModel = ConnectionTreeModel;
 
                 if (Settings.Default.UseSQLServer)
                 {
